requirements.txt
# Video/Audio Processing
moviepy==1.0.3

# Google Cloud APIs
google-cloud-speech==2.23.0
google-cloud-translate==3.14.0
google-cloud-storage==2.14.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
loguru==0.7.2

# Testing (dev)
pytest==7.4.3
pytest-cov==4.1.0
